+++ instrumentation/opentelemetry-instrumentation-asgi/src/opentelemetry/instrumentation/asgi/__init__.py
@@ -64,11 +64,7 @@ carrier_getter = CarrierGetter()
 def collect_request_attributes(scope):
     """Collects HTTP request attributes from the ASGI scope and returns a
     dictionary to be used as span creation attributes."""
-    server = scope.get("server") or ["0.0.0.0", 80]
-    port = server[1]
-    server_host = server[0] + (":" + str(port) if port != 80 else "")
-    full_path = scope.get("root_path", "") + scope.get("path", "")
-    http_url = scope.get("scheme", "http") + "://" + server_host + full_path
+    server_host, port, http_url = get_host_port_url_tuple(scope)
     query_string = scope.get("query_string")
     if query_string and http_url:
         if isinstance(query_string, bytes):
@@ -103,6 +99,17 @@ def collect_request_attributes(scope):
     result = {k: v for k, v in result.items() if v is not None}
 
     return result
+
+
+def get_host_port_url_tuple(scope):
+    """Returns (host, port, full_url) tuple.
+    """
+    server = scope.get("server") or ["0.0.0.0", 80]
+    port = server[1]
+    server_host = server[0] + (":" + str(port) if port != 80 else "")
+    full_path = scope.get("root_path", "") + scope.get("path", "")
+    http_url = scope.get("scheme", "http") + "://" + server_host + full_path
+    return server_host, port, http_url
 
 
 def set_status_code(span, status_code):
@@ -152,12 +159,13 @@ class OpenTelemetryMiddleware:
             Optional: Defaults to get_default_span_details.
     """
 
-    def __init__(self, app, span_details_callback=None):
+    def __init__(self, app, excluded_urls=None, span_details_callback=None):
         self.app = guarantee_single_callable(app)
         self.tracer = trace.get_tracer(__name__, __version__)
         self.span_details_callback = (
             span_details_callback or get_default_span_details
         )
+        self.excluded_urls = excluded_urls
 
     async def __call__(self, scope, receive, send):
         """The ASGI application
@@ -168,6 +176,10 @@ class OpenTelemetryMiddleware:
             send: An awaitable callable taking a single dictionary as argument.
         """
         if scope["type"] not in ("http", "websocket"):
+            return await self.app(scope, receive, send)
+
+        _, _, url = get_host_port_url_tuple(scope)
+        if self.excluded_urls and self.excluded_urls.url_disabled(url):
             return await self.app(scope, receive, send)
 
         token = context.attach(propagators.extract(carrier_getter, scope))