+++ instrumentation/opentelemetry-instrumentation-fastapi/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Added support for excluding some routes with env var `OTEL_PYTHON_FASTAPI_EXCLUDED_URLS`
+  ([#237](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/237))
+
 ## Version 0.11b0
 
 Released 2020-07-28