+++ instrumentation/opentelemetry-instrumentation-starlette/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+- Added support for excluding some routes with env var `OTEL_PYTHON_STARLETTE_EXCLUDED_URLS`
+  ([#237](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/237))
 
 ## Version 0.10b0